Module reth_transaction_pool::blobstore::disk

A simple diskstore for blobs

Structs

DiskFileBlobStore

A blob store that stores blob data on disk.

DiskFileBlobStoreConfig

Configuration for a disk file blob store.

Enums

DiskFileBlobStoreError

Errors that can occur when interacting with a disk file blob store.

OpenDiskFileBlobStore

How to open a disk file blob store.

Constants

DEFAULT_MAX_CACHED_BLOBS

How many BlobTransactionSidecar to cache in memory.