Trait reth_transaction_pool::error::PoolTransactionError

pub trait PoolTransactionError: Error + Send + Sync {
    // Required method
    fn is_bad_transaction(&self) -> bool;
}

A trait for additional errors that can be thrown by the transaction pool.

For example during validation TransactionValidator::validate_transaction

Required Methods

fn is_bad_transaction(&self) -> bool

Returns true if the error was caused by a transaction that is considered bad in the context of the transaction pool and warrants peer penalization.

See PoolError::is_bad_transaction.