Module reth_transaction_pool::test_utils

Common test helpers for mocking a pool Internal helpers for testing.

Structs

MockOrdering

Used to define a specific ordering for transactions, providing a priority value based on the effective tip per gas and base fee of the given MockTransaction.

MockTransactionDistribution

A configured distribution that can generate transactions

MockTransactionFactory

A factory for creating and managing various types of mock transactions.

MockTransactionSet

A set of MockTransactions that can be modified at once

TransactionBuilder

A Builder type to configure and create a transaction.

TransactionGenerator

A generator for transactions for testing purposes

Enums

MockTransaction

A Bare transaction type used for testing.

Functions

mock_tx_pool

Create an empty TxPool

testing_pool

Returns a new Pool used for testing purposes

testing_pool_with_validator

Returns a new Pool used for testing purposes

Type Aliases

MockTxPool

A transaction pool implementation using MockOrdering for transaction ordering.

MockValidTx

A validated transaction in the transaction pool, using MockTransaction as the transaction type.

TestPool

A Pool used for testing