Trait reth_transaction_pool::PoolTransaction

pub trait PoolTransaction: Debug + Send + Sync + FromRecoveredPooledTransaction + FromRecoveredTransaction + IntoRecoveredTransaction {

    // Required methods
    fn hash(&self) -> &TxHash;
    fn sender(&self) -> Address;
    fn nonce(&self) -> u64;
    fn cost(&self) -> U256;
    fn gas_limit(&self) -> u64;
    fn max_fee_per_gas(&self) -> u128;
    fn access_list(&self) -> Option<&AccessList>;
    fn max_priority_fee_per_gas(&self) -> Option<u128>;
    fn max_fee_per_blob_gas(&self) -> Option<u128>;
    fn effective_tip_per_gas(&self, base_fee: u64) -> Option<u128>;
    fn priority_fee_or_price(&self) -> u128;
    fn kind(&self) -> &TransactionKind;
    fn input(&self) -> &[u8] ;
    fn size(&self) -> usize;
    fn tx_type(&self) -> u8;
    fn encoded_length(&self) -> usize;
    fn chain_id(&self) -> Option<u64>;

    // Provided methods
    fn to(&self) -> Option<Address> { ... }
    fn is_eip1559(&self) -> bool { ... }
    fn is_eip4844(&self) -> bool { ... }

}

Trait for transaction types used inside the pool

Required Methods

fn hash(&self) -> &TxHash

Hash of the transaction.

fn sender(&self) -> Address

The Sender of the transaction.

fn nonce(&self) -> u64

Returns the nonce for this transaction.

fn cost(&self) -> U256

Returns the cost that this transaction is allowed to consume:

For EIP-1559 transactions: max_fee_per_gas * gas_limit + tx_value. For legacy transactions: gas_price * gas_limit + tx_value. For EIP-4844 blob transactions: max_fee_per_gas * gas_limit + tx_value + max_blob_fee_per_gas * blob_gas_used.

fn gas_limit(&self) -> u64

Amount of gas that should be used in executing this transaction. This is paid up-front.

fn max_fee_per_gas(&self) -> u128

Returns the EIP-1559 the maximum fee per gas the caller is willing to pay.

For legacy transactions this is gas_price.

This is also commonly referred to as the “Gas Fee Cap” (GasFeeCap).

fn access_list(&self) -> Option<&AccessList>

Returns the access_list for the particular transaction type. For Legacy transactions, returns default.

fn max_priority_fee_per_gas(&self) -> Option<u128>

Returns the EIP-1559 Priority fee the caller is paying to the block author.

This will return None for non-EIP1559 transactions

fn max_fee_per_blob_gas(&self) -> Option<u128>

Returns the EIP-4844 max fee per data gas

This will return None for non-EIP4844 transactions

fn effective_tip_per_gas(&self, base_fee: u64) -> Option<u128>

Returns the effective tip for this transaction.

For EIP-1559 transactions: min(max_fee_per_gas - base_fee, max_priority_fee_per_gas). For legacy transactions: gas_price - base_fee.

fn priority_fee_or_price(&self) -> u128

Returns the max priority fee per gas if the transaction is an EIP-1559 transaction, and otherwise returns the gas price.

fn kind(&self) -> &TransactionKind

Returns the transaction’s TransactionKind, which is the address of the recipient or TransactionKind::Create if the transaction is a contract creation.

fn input(&self) -> &[u8]

Returns the input data of this transaction.

fn size(&self) -> usize

Returns a measurement of the heap usage of this type and all its internals.

fn tx_type(&self) -> u8

Returns the transaction type

fn encoded_length(&self) -> usize

Returns the length of the rlp encoded transaction object

Note: Implementations should cache this value.

fn chain_id(&self) -> Option<u64>

Returns chain_id

Provided Methods

fn to(&self) -> Option<Address>

Returns the recipient of the transaction if it is not a TransactionKind::Create transaction.

fn is_eip1559(&self) -> bool

Returns true if the transaction is an EIP-1559 transaction.

fn is_eip4844(&self) -> bool

Returns true if the transaction is an EIP-4844 transaction.

Object Safety

This trait is not object safe.