Trait reth_transaction_pool::TransactionPoolExt

pub trait TransactionPoolExt: TransactionPool {
    // Required methods
    fn set_block_info(&self, info: BlockInfo);
    fn on_canonical_state_change(&self, update: CanonicalStateUpdate<'_>);
    fn update_accounts(&self, accounts: Vec<ChangedAccount>);
    fn delete_blob(&self, tx: B256);
    fn delete_blobs(&self, txs: Vec<B256>);
}

Extension for TransactionPool trait that allows to set the current block info.

Required Methods

fn set_block_info(&self, info: BlockInfo)

Sets the current block info for the pool.

fn on_canonical_state_change(&self, update: CanonicalStateUpdate<'_>)

Event listener for when the pool needs to be updated

Implementers need to update the pool accordingly. For example the base fee of the pending block is determined after a block is mined which affects the dynamic fee requirement of pending transactions in the pool.

fn update_accounts(&self, accounts: Vec<ChangedAccount>)

Updates the accounts in the pool

fn delete_blob(&self, tx: B256)

Deletes the blob sidecar for the given transaction from the blob store

fn delete_blobs(&self, txs: Vec<B256>)

Deletes multiple blob sidecars from the blob store

Object Safety

This trait is not object safe.