Trait reth_transaction_pool::validate::TransactionValidator

pub trait TransactionValidator: Send + Sync {
    type Transaction: PoolTransaction;

    // Required method
    fn validate_transaction<'life0, 'async_trait>(
        &'life0 self,
        origin: TransactionOrigin,
        transaction: Self::Transaction
    ) -> Pin<Box<dyn Future<Output = TransactionValidationOutcome<Self::Transaction>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn validate_transactions<'life0, 'async_trait>(
        &'life0 self,
        transactions: Vec<(TransactionOrigin, Self::Transaction)>
    ) -> Pin<Box<dyn Future<Output = Vec<TransactionValidationOutcome<Self::Transaction>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn on_new_head_block(&self, _new_tip_block: &SealedBlock) { ... }
}

Provides support for validating transaction at any given state of the chain

Required Associated Types

type Transaction: PoolTransaction

The transaction type to validate.

Required Methods

fn validate_transaction<'life0, 'async_trait>( &'life0 self, origin: TransactionOrigin, transaction: Self::Transaction ) -> Pin<Box<dyn Future<Output = TransactionValidationOutcome<Self::Transaction>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Validates the transaction and returns a TransactionValidationOutcome describing the validity of the given transaction.

This will be used by the transaction-pool to check whether the transaction should be inserted into the pool or discarded right away.

Implementers of this trait must ensure that the transaction is well-formed, i.e. that it complies at least all static constraints, which includes checking for:

chain id

gas limit

max cost

nonce >= next nonce of the sender

…

See InvalidTransactionError for common errors variants.

The transaction pool makes no additional assumptions about the validity of the transaction at the time of this call before it inserts it into the pool. However, the validity of this transaction is still subject to future (dynamic) changes enforced by the pool, for example nonce or balance changes. Hence, any validation checks must be applied in this function.

See TransactionValidationTaskExecutor for a reference implementation.

Provided Methods

fn validate_transactions<'life0, 'async_trait>( &'life0 self, transactions: Vec<(TransactionOrigin, Self::Transaction)> ) -> Pin<Box<dyn Future<Output = Vec<TransactionValidationOutcome<Self::Transaction>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Validates a batch of transactions.

Must return all outcomes for the given transactions in the same order.

See also Self::validate_transaction.

fn on_new_head_block(&self, _new_tip_block: &SealedBlock)

Invoked when the head block changes.

This can be used to update fork specific values (timestamp).