Trait reth_transaction_pool::BestTransactions

pub trait BestTransactions: Iterator + Send {
    // Required methods
    fn mark_invalid(&mut self, transaction: &Self::Item);
    fn no_updates(&mut self);
    fn skip_blobs(&mut self);
    fn set_skip_blobs(&mut self, skip_blobs: bool);

    // Provided method
    fn filter<P>(self, predicate: P) -> BestTransactionFilter<Self, P> 
       where P: FnMut(&Self::Item) -> bool,
             Self: Sized { ... }
}

An Iterator that only returns transactions that are ready to be executed.

This makes no assumptions about the order of the transactions, but expects that all transactions are valid (no nonce gaps.) for the tracked state of the pool.

Note: this iterator will always return the best transaction that it currently knows. There is no guarantee transactions will be returned sequentially in decreasing priority order.

Required Methods

fn mark_invalid(&mut self, transaction: &Self::Item)

Mark the transaction as invalid.

Implementers must ensure all subsequent transaction don’t depend on this transaction. In other words, this must remove the given transaction and drain all transaction that depend on it.

fn no_updates(&mut self)

An iterator may be able to receive additional pending transactions that weren’t present it the pool when it was created.

This ensures that iterator will return the best transaction that it currently knows and not listen to pool updates.

fn skip_blobs(&mut self)

Skip all blob transactions.

There’s only limited blob space available in a block, once exhausted, EIP-4844 transactions can no longer be included.

If called then the iterator will no longer yield blob transactions.

Note: this will also exclude any transactions that depend on blob transactions.

fn set_skip_blobs(&mut self, skip_blobs: bool)

Controls whether the iterator skips blob transactions or not.

If set to true, no blob transactions will be returned.

Provided Methods

fn filter<P>(self, predicate: P) -> BestTransactionFilter<Self, P>
where P: FnMut(&Self::Item) -> bool, Self: Sized,

Creates an iterator which uses a closure to determine if a transaction should be yielded.

Given an element the closure must return true or false. The returned iterator will yield only the elements for which the closure returns true.

Descendant transactions will be skipped.