Trait reth_transaction_pool::EthPoolTransaction

pub trait EthPoolTransaction: PoolTransaction {
    // Required methods
    fn take_blob(&mut self) -> EthBlobTransactionSidecar;
    fn as_eip4844(&self) -> Option<&TxEip4844>;
    fn validate_blob(
        &self,
        blob: &BlobTransactionSidecar,
        settings: &KzgSettings
    ) -> Result<(), BlobTransactionValidationError>;

    // Provided method
    fn blob_count(&self) -> usize { ... }
}

An extension trait that provides additional interfaces for the EthTransactionValidator.

Required Methods

fn take_blob(&mut self) -> EthBlobTransactionSidecar

Extracts the blob sidecar from the transaction.

fn as_eip4844(&self) -> Option<&TxEip4844>

Returns the transaction as EIP-4844 transaction if it is one.

fn validate_blob( &self, blob: &BlobTransactionSidecar, settings: &KzgSettings ) -> Result<(), BlobTransactionValidationError>

Validates the blob sidecar of the transaction with the given settings.

Provided Methods

fn blob_count(&self) -> usize

Returns the number of blobs this transaction has.

Object Safety

This trait is not object safe.