Module reth_transaction_pool::validate

Transaction validation abstractions.

Structs

EthTransactionValidator

A TransactionValidator implementation that validates ethereum transaction. Validator for Ethereum transactions.

EthTransactionValidatorBuilder

A TransactionValidator implementation that validates ethereum transaction. A builder for TransactionValidationTaskExecutor

TransactionValidationTaskExecutor

A spawnable task that performs transaction validation. A TransactionValidator implementation that validates ethereum transaction.

ValidPoolTransaction

A valid transaction in the pool.

ValidationTask

A spawnable task that performs transaction validation. A service that performs validation jobs.

Enums

TransactionValidationOutcome

A Result type returned after checking a transaction’s validity.

TransactionValidatorError

Validation Errors that can occur during transaction validation.

ValidTransaction

A wrapper type for a transaction that is valid and has an optional extracted EIP-4844 blob transaction sidecar.

Constants

MAX_CODE_BYTE_SIZE

Validation constants. Maximum bytecode to permit for a contract.

MAX_INIT_CODE_BYTE_SIZE

Validation constants. Maximum initcode to permit in a creation transaction and create instructions.

MAX_TX_INPUT_BYTES

Validation constants. MAX_TX_INPUT_BYTES is the maximum size a single transaction can have. This field has non-trivial consequences: larger transactions are significantly harder and more expensive to propagate; larger transactions also take more resources to validate whether they fit into the pool or not.

TX_SLOT_BYTE_SIZE

Validation constants. TX_SLOT_BYTE_SIZE is used to calculate how many data slots a single transaction takes up based on its byte size. The slots are used as DoS protection, ensuring that validating a new transaction remains a constant operation (in reality O(maxslots), where max slots are 4 currently).

Traits

TransactionValidator

Provides support for validating transaction at any given state of the chain

Functions

ensure_intrinsic_gas

A TransactionValidator implementation that validates ethereum transaction. Ensures that gas limit of the transaction exceeds the intrinsic gas of the transaction.

ensure_max_init_code_size

A TransactionValidator implementation that validates ethereum transaction. Ensure that the code size is not greater than max_init_code_size. max_init_code_size should be configurable so this will take it as an argument.